requirements.txt
requests
phonenumbers
selenium
beautifulsoup4
urllib3
dnsresolver
whois
